Refactor package and file structure
